Organize FindClassPath into modules
